extract createGCEInstance, generateGCEResources
Extrated `createGCEInstance` and `generateGCEResources` from
`deploy.js`.
