+++ src/commands/deploy.js
@@ -13,9 +13,10 @@ const { validateIntegrationName } = require('../gcp/validateIntegrationName');
 const createBuckets = require('../gcp/createBuckets');
 const { createCloudFunction } = require('../gcp/createCloudFunction');
 const addIntegrationToDirectory = require('../utils/addIntegrationToDirectory');
+
 // Mat moved functions outside of deploy
 const yaml = require('js-yaml');
-const createEmbulkConfig = require('../utils/createEmbulkConfig');
+const createGCEInstance = require('../gcp/createGCEInstance');
 
 module.exports = async (args) => {
   const mashrConfigObj = await readYaml('./mashr_config.yml');
@@ -51,127 +52,4 @@ module.exports = async (args) => {
   //  overwrite?
   //  - if default region doesn't exist in init, where is the bucket, GCE, and
   //  GCF created? Does it matter?
-};
-
-const { readFile } = require('../utils/fileUtils');
-const generateGCEResources = async (mashrConfigObj) => {
-  const dockerfile = await readFile(`${__dirname}/../../templates/docker/Dockerfile`);
-  const gemInstallationScript = createGemInstallationScript(mashrConfigObj.mashr.embulk_gems);
-  const keyfile = await readFile(`${mashrConfigObj.mashr.json_keyfile}`);
-  const embulkScript = createEmbulkScript(mashrConfigObj.mashr.embulk_run_command);
-  const crontab = await readFile(`${__dirname}/../../templates/docker/crontab`);
-  const embulkConfig = createEmbulkConfig(mashrConfigObj);
-
-  return {
-    dockerfile,
-    gemInstallationScript,
-    keyfile,
-    embulkScript,
-    crontab,
-    embulkConfig,
-  };
-};
-
-
-const createEmbulkScript = (runCommand) => {
-  // TODO: place logs in stackdriver
-  // TODO: what to do with logs? Does the log file get too large?
-  // diff file run from root of container. Can't use it after?
-  runCommand = runCommand.replace(
-    'embulk_config.yml', '/root/mashr/embulk_config.yml.liquid');
-
-  const script =
-`#!/bin/bash
-export DATE=$(date +"%Y-%m-%dT%H-%M-%S-%3N")
-
-${runCommand} >> /var/log/cron.log 2>&1
-`;
-
-  return script;
-}
-
-const createGemInstallationScript = (gems) => {
-  if (!gems) return '#!/bin/bash';
-
-  const installGemsArray = gems.map((name) => (
-      `embulk gem install ${name}`
-    )
-  );
-
-  return `#!/bin/bash\n${installGemsArray.join('\n')}`;
-};
-
-const Compute = require('@google-cloud/compute');
-const createGCEInstance = async (mashrConfigObj) => {
-
-  const compute = new Compute();
-
-  const zone = compute.zone('us-central1-a');
-
-  const {
-    dockerfile,
-    gemInstallationScript,
-    embulkConfig,
-    keyfile,
-    embulkScript,
-    crontab,
-  } = await generateGCEResources(mashrConfigObj);
-
-  const config = {
-    os: 'debian-9',
-    http: true,
-    machineType: 'g1-small',
-    tags: ["mashr"],
-    metadata: {
-      items: [
-        {
-          key: 'startup-script',
-          value: `#! /bin/bash
-
-          sudo apt-get update
-          sudo apt install apt-transport-https ca-certificates curl gnupg2 software-properties-common -y
-          curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
-          sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/debian $(lsb_release -cs) stable"
-          sudo apt update
-          sudo apt install docker-ce -y
-
-          cd /
-          sudo mkdir app
-          cd app
-          echo "${dockerfile.toString()}" > Dockerfile
-          echo '${embulkScript}' > embulkScript.sh
-          echo '${crontab.toString()}' > crontab
-          sudo mkdir mashr
-          echo "${gemInstallationScript}" > mashr/install_gems.sh
-          echo "${embulkConfig}" > mashr/embulk_config.yml.liquid
-          printf "%s\n" '${keyfile.toString()}' > mashr/keyfile.json
-
-          sudo docker pull jacobleecd/mashr:latest
-          sudo docker build -t mashr .
-          sudo docker run -d -v /mashr --name embulk-container mashr
-          `
-        },
-      ],
-    },
-  };
-
-  const vm = zone.vm(mashrConfigObj.mashr.integration_name);
-  // const vm = zone.vm('newname');
-
-  vm.create(config);
-  // vm.create(config, function(err, vm, operation, apiResponse) {
-  //   // `vm` is a VM object.
-
-  //   // `operation` is an Operation object that can be used to check the
-  //   // status of the request.
-  //   console.log('!!!!!!!!!!!!!!');
-  //   console.log('VM: ', vm);
-  //   console.log('!!!!!!!!!!!!!!');
-  //   console.log('operation: ', operation);
-  //   console.log('!!!!!!!!!!!!!!');
-  //   console.log('apiResponse: ', apiResponse);
-  //   console.log('!!!!!!!!!!!!!!');
-  //   console.log('error: ', err);
-  // });
-
 };
